Delete prtype

+++ mvc/views/adminpages/listproducttype.php
@@ -35,6 +35,9 @@ The above copyright notice and this permission notice shall be included in all c
 </head>
 
 <body class="">
+    <?php
+        include 'ConfirmDelete.php';
+    ?>
     <div class="wrapper ">
         <div class="sidebar" style="background-color: #f5e0c2;">
             <!--
@@ -140,7 +143,7 @@ The above copyright notice and this permission notice shall be included in all c
                                                     echo $row["TenLSP"] .
                                                         "</td>
                                                          <td style='text-align: center; width: 200px;'>
-                                                        <a href='./deleteprtype'><button class='snip0059'><span class='material-icons' style='font-size: 15px;'>
+                                                        ".'<a href="./deleteprtype?MaLSP='.$row[0].'"'."><button onclick='return confirmSubmit()' class='snip0059'><span class='material-icons' style='font-size: 15px;'>
                                                             delete
                                                             </span></button></a>
                                                         ".'<a href="./editprtype?MaLSP='.$row[0].'?TenLSP='.$row[1].'"'."><button class='snip0059'><span class='material-icons' style='font-size: 15px;'>
